refactor trpc folder
