+++ src/libs/trpc/middleware/userAuth.test.ts
@@ -4,7 +4,7 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { createCallerFactory } from '@/libs/trpc/lambda';
 import { AuthContext, createContextInner } from '@/libs/trpc/lambda/context';
 
-import { trpc } from '../init';
+import { trpc } from '../lambda/init';
 import { userAuth } from './userAuth';
 
 const appRouter = trpc.router({
